+++ CHANGES.md
@@ -1,3 +1,6 @@
 
+## Development
+- add `set_all` and `clear_all`
+
 ## 0.0.2
 - Initial release